teacher portal add lecture done

+++ database/migrations/2020_02_16_132402_create_employees_table.php
@@ -25,9 +25,9 @@ class CreateEmployeesTable extends Migration
             $table->string('employee_dob')->nullable();
             $table->string('marital_status')->nullable();
             
-            $table->integer('branch_id')->nullable();
-            $table->integer('user_id')->nullable();
-            $table->integer('dept_id')->nullable();
+            $table->bigInteger('branch_id')->nullable();
+            $table->bigInteger('user_id')->nullable();
+            $table->bigInteger('dept_id')->nullable();
             $table->timestamps();
         });
     }
